+++ @navikt/core/react/src/form/textfield/text-field.stories.tsx
@@ -1,6 +1,6 @@
 import { Meta, StoryObj } from "@storybook/react";
 import React from "react";
-import { TextField } from "../index";
+import TextField from "./TextField";
 
 export default {
   title: "ds-react/TextField",