+++ @navikt/core/react/src/layout/box/index.ts
@@ -1,1 +1,2 @@
-export { Box, type BoxProps } from "./Box";
+"use client";
+export { default as Box, type BoxProps } from "./Box";