+++ @navikt/core/react/src/layout/stack/Stack.tsx
@@ -76,3 +76,5 @@ export const Stack: OverridableComponent<StackProps, HTMLDivElement> =
       );
     },
   );
+
+export default Stack;