+++ @navikt/core/react/src/layout/stack/index.ts
@@ -1,4 +1,5 @@
-export { HStack, type HStackProps } from "./HStack";
-export { VStack, type VStackProps } from "./VStack";
-export { Stack, type StackProps } from "./Stack";
-export { Spacer } from "./Spacer";
+"use client";
+export { default as HStack, type HStackProps } from "./HStack";
+export { default as VStack, type VStackProps } from "./VStack";
+export { default as Stack, type StackProps } from "./Stack";
+export { default as Spacer } from "./Spacer";