+++ @navikt/core/react/src/link-panel/index.ts
@@ -1,2 +1,10 @@
-export { default as LinkPanel } from "./LinkPanel";
-export { type LinkPanelProps } from "./LinkPanel";
+"use client";
+export { default as LinkPanel, type LinkPanelProps } from "./LinkPanel";
+export {
+  default as LinkPanelTitle,
+  type LinkPanelTitleProps,
+} from "./LinkPanelTitle";
+export {
+  default as LinkPanelDescription,
+  type LinkPanelDescriptionProps,
+} from "./LinkPanelDescription";