+++ @navikt/core/react/src/table/stories/table-1.stories.tsx
@@ -1,7 +1,7 @@
 import React, { useState } from "react";
 import { Table, TableProps } from "../";
 import { Button } from "../../button";
-import { Checkbox } from "../../form";
+import { Checkbox } from "../../form/checkbox";
 import { VStack } from "../../layout/stack";
 import {
   Expandable,