+++ @navikt/core/react/src/util/hooks/useClientLayoutEffect.ts
@@ -1,3 +1,5 @@
+"use client";
+
 import { useLayoutEffect } from "react";
 
 export const useClientLayoutEffect = globalThis?.document