+++ @navikt/core/react/src/util/index.ts
@@ -1,6 +1,6 @@
-export type * from "./types";
+export type { OverridableComponent } from "./types";
 export { default as debounce } from "./debounce";
-export * from "./omit";
+export { omit } from "./omit";
 export { useClientLayoutEffect } from "./hooks/useClientLayoutEffect";
 export { useEventListener } from "./hooks/useEventListener";
 export { useId } from "./hooks/useId";